Restore pre-#27436 BCG core; adapt diffusion side to it
The srt breakable CUDA graph core returns byte-identical to its state
before sgl-project/sglang#27436 (back at
srt/model_executor/runner_backend_utils/breakable_cuda_graph, dedup
support and capture-failure logging restored, top-level
srt/breakable_cuda_graph package removed).

The diffusion BCG feature is kept and adapted to the pre-PR core:
- replay-token machinery moves to
  multimodal_gen/runtime/breakable_cuda_graph/replay_token.py; the mm
  runner stamps a token around each graph.replay()
- tuple-returning attention forwards (UlyssesAttention, used by
  Z-Image) are boxed into an attribute object for the break point,
  since the pre-PR _copy_output copies objects-with-tensor-attrs in
  place but ignores tuples
- layer.py imports repoint accordingly

Verified: core files byte-identical to ead1e490b; all consumers
import; GPU capture/replay of a boxed tuple break point produces
fresh outputs on every replay; 16/16 BCG unit tests pass.
